+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [9.6.0] - Unreleased
+## [9.6.0] - 2020-12-30
 
 ### Changed
 
 - MarkupError exception raise from None to omit internal exception
 - Factored out RichHandler.render and RichHandler.render_message for easier extending
+- Display pretty printed value in rich.inspect
 
 ### Added